SEMICOLONSgit statusgit statusgit status

+++ app/assets/javascripts/carts.js
@@ -7,39 +7,43 @@ function add_to_cart(cart_id, part_id) {
 		// Variables
 		var cur = "";
 		var max_inv = parseInt(data["on_hand"]);
-		// create the HTML of the table entry
-		html = "<tr class = 'borders fit' id='part_"+ part_id + "'>" +
-					"<td id='c5'><button type='button' class = 'btn-default btn-s' id = 'small' onclick = 'delete_item("+ cart_id +","+ part_id +")'>X</button></td>" + 
-					"<td id='c6'>" + data["part_number"] + "</td>" +
-					"<td id='c6'>" + data["price"] + "</td>" +
-					"<td id='c8'>" + data["description"] + "</td>" +
-					"<td id='c9'><button type=button' class = 'btn-default btn-xs' onclick='add_to_cart("+ cart_id +","+ part_id +")'>+</button></td>" +	
-					"<td class='amount' id='c10'><a href='#basicModal' role='button' class='btn-default btn-s' data-toggle='modal' onclick='call_modal("+ part_id +")'><b>1</a></b></td>" + 
-					"<td id='c11'><button type=button' class = 'btn-default btn-xs' onclick='remove_from_cart("+ cart_id +","+ part_id +")'>-</button></td>" +	
-			   "</tr>";
-		// If the entry is in the cart
-		if($(".cart table #part_" + part_id).length) {
-			// update the count of the part
-			cur = $(".cart table #part_" + part_id + " .amount").text();
-			if (max_inv > cur) {
-				amount = parseInt(cur) + 1;
-				console.log(amount);
-				$.post("update_part.json", {cart_id: cart_id, part_id: part_id, quantity_requested: amount}).done(function(){
-					$(".cart table #part_" + part_id + " .amount").html("<a href='#basicModal' role='button' class='btn-default btn-s' data-toggle='modal'><b>"+amount+"</a></b>");
+		// Let's make sure there is inventory
+		if (max_inv > 0) {
+			// create the HTML of the table entry
+			html = "<tr class = 'borders fit' id='part_"+ part_id + "'>" +
+						"<td id='c5'><button type='button' class = 'btn-default btn-s' id = 'small' onclick = 'delete_item("+ cart_id +","+ part_id +")'>X</button></td>" + 
+						"<td id='c6'>" + data["part_number"] + "</td>" +
+						"<td id='c6'>" + data["price"] + "</td>" +
+						"<td id='c8'>" + data["description"] + "</td>" +
+						"<td id='c9'><button type=button' class = 'btn-default btn-xs' onclick='add_to_cart("+ cart_id +","+ part_id +")'>+</button></td>" +	
+						"<td class='amount' id='c10'><a href='#basicModal' role='button' class='btn-default btn-s' data-toggle='modal' onclick='call_modal("+ part_id +")'><b>1</a></b></td>" + 
+						"<td id='c11'><button type=button' class = 'btn-default btn-xs' onclick='remove_from_cart("+ cart_id +","+ part_id +")'>-</button></td>" +	
+				   "</tr>";
+			// If the entry is in the cart
+			if($(".cart table #part_" + part_id).length) {
+				// update the count of the part
+				cur = $(".cart table #part_" + part_id + " .amount").text();
+				if (max_inv > cur) {
+					amount = parseInt(cur) + 1;
+					console.log(amount);
+					$.post("update_part.json", {cart_id: cart_id, part_id: part_id, quantity_requested: amount}).done(function(){
+						$(".cart table #part_" + part_id + " .amount").html("<a href='#basicModal' role='button' class='btn-default btn-s' data-toggle='modal'><b>"+amount+"</a></b>");
+						get_total(cart_id);
+					});
+				}
+			}
+			// If it is not in the cart 
+			else {
+				$.post("add_part.json", {cart_id: cart_id, part_id:part_id}).done(function(data){
+					// add the new entry to the carts
+					$(".cart table tr:last").after("" + html);
 					get_total(cart_id);
 				});
 			}
 		}
-		// If it is not in the cart 
-		else {
-			$.post("add_part.json", {cart_id: cart_id, part_id:part_id}).done(function(data){
-				// add the new entry to the carts
-				$(".cart table tr:last").after("" + html);
-				get_total(cart_id);
-			});
-		}
 	});
 	get_total(cart_id);
+	
 }
 
 // Remove a part entry from the cart
@@ -51,7 +55,7 @@ function remove_from_cart(cart_id, part_id) {
 		// If it is in the cart ger the current amount
 		cur = $(".cart table #part_" + part_id + " .amount").text();
 		amount = parseInt(cur) - 1;
-		console.log("Value: " +  amount)
+		console.log("Value: " +  amount);
 		// If the current amount is more than 1
 		if(amount != 0) {
 			$.post("update_part.json", {cart_id: cart_id, part_id: part_id, quantity_requested: amount}).done(function(){
@@ -72,8 +76,10 @@ function remove_from_cart(cart_id, part_id) {
 
 function delete_item(cart_id, part_id){
 	$.post("remove_part.json", {cart_id: cart_id, part_id:part_id}).done(function(data) {
-				$(".cart table #part_" + part_id).remove();	
-			});
+		$(".cart table #part_" + part_id).remove();	
+		get_total(cart_id);
+	});
+	get_total(cart_id);
 }
 
 function add_quantity_manually(cart_id){
@@ -95,6 +101,7 @@ function add_quantity_manually(cart_id){
 				get_total(cart_id);
 			});
 	}
+	get_total(cart_id);
 }
 
 function call_modal(part_id){
